Added mkSpiffs function when MainForm opens.
The program will run mkSpiffs if there is a table.txt file in the 'image' directory in the firmware folder. It creates a lookup.bin file if it does not yet exists, or overwrites the existing file. If there is no table.txt file, then the program uses the current lookup.bin file instead of writing a new one.